add dinov2 ltdetr
